Fix u128 timestamp/duration fields for msgpack compatibility
rmp-serde encodes u128 as a 16-byte binary blob, breaking the JS side
which expects numbers. Change latency_ns, engine_ns, ts_ms/ts_ns
timestamps and ping/pong timestamps from u128 to u64 in the wire types.
order_id stays u128 since it's a UUID and never read numerically by
SimEngine. JSON path is unaffected (u64 still serializes as a number).

Verified via msgpack WS round-trip: latency_ns=18708, engine_ns=43375
land as numbers; latency stream (JSON) and bot driver still work.

+++ hftx/exchange-service/src/exchange.rs
@@ -92,7 +92,7 @@ impl Exchange {
             best_ask: orderbook.best_ask(),  // Lowest ask price
             bid_levels,
             ask_levels,
-            last_update: SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos(),
+            last_update: SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos() as u64,
         })
     }
 
@@ -149,7 +149,7 @@ impl Exchange {
             symbol: symbol.to_string(),
             bids,
             asks,
-            timestamp: SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos(),
+            timestamp: SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_nanos() as u64,
         })
     }
